+++ CHANGELOG.md
@@ -1,1 +1,10 @@
-TODO
+## 3.0
+- updated for Minecraft 1.4.2 (protocol version 47)
+- added Forge compatibility layer
+
+## 2.0
+- updated for Minecraft 1.3.1 (protocol version 39)
+- added support for inspecting packets used by integrated singleplayer server
+
+## 1.0
+- initial release for Minecraft 1.2.5 (protocol version 29)